Incorporate an inner change score into recommender sort function
